Implement search survey for global search.
